bot/requirements.txt
discord.py==2.6.4  # SECURITY: Updated from 2.3.2 for latest security patches
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0  # Natural language date parsing for event scheduling
dateparser==1.2.0  # Advanced date parsing for visualizations
aiohttp==3.13.3  # SECURITY: Updated from 3.9.5 for CVE fixes (HTTP request smuggling, etc.)
tavily-python==0.3.9  # Updated for latest features
requests==2.32.5  # SECURITY: Updated from 2.32.3 for latest CVE patches
httpx==0.28.1  # SECURITY: Updated from 0.24.1 for security fixes
openai==1.109.1  # OpenAI API for embeddings (RAG system) - latest v1.x stable (v2.x has breaking changes)
pgvector==0.3.5  # PostgreSQL vector extension Python client
scikit-learn==1.5.1  # Updated for performance and security
networkx==3.3  # Updated for bug fixes
pandas==2.2.2  # Updated for performance improvements
numpy==1.26.4  # Keep on 1.x branch (numpy 2.x has breaking changes)
cryptography==46.0.3  # SECURITY: Critical update from 42.0.8 for OpenSSL bindings and CVE fixes
Pillow==11.0.0  # SECURITY: Updated from 10.3.0, staying on 11.x for compatibility (12.x may have breaking changes)
matplotlib==3.9.0  # Visualization library for charts and tables
seaborn==0.13.2  # Statistical visualizations built on matplotlib
llmlingua==0.2.2  # Prompt compression for reducing context token usage
redis==5.0.1  # Redis client for in-memory caching
cachetools==5.3.3  # In-memory caching with TTL support (used by iRacing meta analyzer)
tenacity==8.2.3  # Retry logic with exponential backoff (used by iRacing API client)
yfinance==0.2.37  # Stock/crypto price lookups
curl_cffi==0.7.4  # Required by yfinance to bypass Yahoo Finance bot detection
beautifulsoup4==4.12.3  # HTML parsing for URL previews
pytz==2024.1  # Timezone handling
yt-dlp==2024.12.6  # YouTube video downloading for frame extraction
youtube-transcript-api==0.6.2  # YouTube captions/transcript extraction
duckduckgo-search==7.0.1  # Image search with safe search support
